requirements.txt
absl-py==1.4.0
amqp==5.1.1
ansicon==1.89.0
APScheduler==3.10.1
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
blessed==1.20.0
cachetools==5.3.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
Django==4.2.1
django-apscheduler==0.6.2
django-cors-headers==3.14.0
django-crontab==0.7.1
django-picklefield==3.1
django-q==1.3.9
django-rest-framework==0.1.0
django-windows-tools==0.2.1
djangorestframework==3.14.0
executing==1.2.0
flatbuffers==23.5.9
gast==0.4.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
h5py==3.8.0
httplib2==0.22.0
idna==3.4
ipykernel==6.22.0
ipython==8.11.0
jax==0.4.10
jedi==0.18.2
jinxed==1.2.0
joblib==1.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
keras==2.12.0
kombu==5.2.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
ml-dtypes==0.1.0
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.1.1
prompt-toolkit==3.0.38
protobuf==4.23.1
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pywin32==305
pyzmq==25.0.2
redis==3.5.3
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post5
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-addons==0.20.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.9.0
typeguard==2.13.3
typing_extensions==4.6.2
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.4
wrapt==1.14.1
